refactored python testing script into smaller functions for ease of debugged and tested in 1.7.1

+++ transformers/how_to_test_from_py_client.py
@@ -1,117 +1,119 @@
-"""Functions to ease testing a new custom transformer from the python client"""
+"""Testing a BYOR Transformer the PyClient - works on 1.7.0 & 1.7.1-17"""
 import pandas as pd
+from h2oai_client import Client
+import sys
+import zipfile
 import os
 import shutil
-import sys
-from h2oai_client import Client
 
+# TODO: re-write the already uploaded data check to account for numpy warning of type mismatch
+import warnings
+warnings.simplefilter(action='ignore', category=FutureWarning)
+
+# Print and Debug Nicely
 pd.set_option('display.max_rows', 50)
 pd.set_option('display.max_columns', 500)
 pd.set_option('display.width', 1000)
 
+# The following are parameters that need to be set to run these functions
+# TODO: to redo this is a nicer way
 
-def connect():
-    # Login info
-    dai_url = "http://IPADDRESS:12345"
-    dai_user = "UserName"
-    dai_pwd = "Password"
+# Connect to Driverless AI
+h2oai = Client('http://18.206.38.108:12345', 'h2oai', 'i-0e5001d094cdc2c90')
 
-    return Client(dai_url, dai_user, dai_pwd)
+# Data Information
+data_file_name = "testdata.csv"
+data_file_location = "/Users/mtanco/Downloads/" + data_file_name
+y = "y"
+
+# Transformers Information
+transformer_name = "MyLogTransformer"
+transformer_file_name = "log_transformer.py"
+transformer_file_location = "/Users/mtanco/Documents/GitProjects/driverlessai-recipes/transformers/numeric/" + transformer_file_name
+
+# Location to Download Files
+download_file_location = "/Users/mtanco/Downloads/"
 
 
-def print_system_custom_transformers():
-    h2oai = connect()
-
+# Print the default & custom transformers on the system, return list of all transformers
+def get_transformers(print_details=True):
     all_transformers = h2oai.list_transformers()
-
-    print(all_transformers[0].dump())
 
     names = list(map(lambda x: x.name, all_transformers))
     types = list(map(lambda x: x.is_custom, all_transformers))
 
-    all_trans = pd.DataFrame(
-        {'name': names,
+    all_trans = pd.DataFrame({
+         'name': names,
          'is_custom': types
-         })
+    })
 
-    print(list(all_trans[all_trans["is_custom"]]["name"]))
+    if print_details:
+        print("GET TRANSFORMERS: ")
+        print("\tCustom Transformers:", list(all_trans[all_trans["is_custom"]]["name"]))
+        print("\tDefault Transformers:", list(all_trans[~all_trans["is_custom"]]["name"]))
+        print("")
+
+    return list(all_trans["name"])
 
 
-def test_debug_pyclient():
-    # Data Information
-    data_file_name = "****.csv"
-    y = "****"
+# Load the custom transformer, exit gracefully if it fails
+# TODO: return error message or logs or if it fails
+def load_transformer(print_details=True):
 
-    # Transformers Information
-    transformer_file_name = "****.py"
+    my_transformer = h2oai.upload_custom_recipe_sync(transformer_file_location)
 
-    # Use empty lists if you want to test how the transformer does in relation to other transformers
-    transformers_noncustom = []
-    transformers_custom_nontesting = []
+    # returns true or false - exit if fails - check DAI UI for error message (make new experiment & upload)
+    if my_transformer:
+        if print_details:
+            print("LOAD TRANSFORMER:")
+            print("\tTransformer uploaded successfully")
+            print("")
+    else:
+        print("LOAD TRANSFORMER:")
+        print("\tTransformer uploaded failed, exiting program.")
+        sys.exit()
 
-    # All Official Transformers
-    transformers_noncustom = ['CVCatNumEncodeTransformer', 'CVTargetEncodeTransformer'
-        , 'CatOriginalTransformer', 'ClusterDistTransformer'
-        , 'ClusterIdTransformer', 'ClusterTETransformer', 'DatesTransformer'
-        , 'EwmaLagsTransformer', 'FrequentTransformer', 'InteractionsTransformer'
-        , 'IsHolidayTransformer', 'LagsAggregatesTransformer', 'LagsInteractionTransformer'
-        , 'LagsTransformer', 'LexiLabelEncoderTransformer', 'NumCatTETransformer', 'NumToCatTETransformer'
-        , 'NumToCatWoEMonotonicTransformer', 'NumToCatWoETransformer', 'OneHotEncodingTransformer'
-        , 'OriginalTransformer', 'SortedLETransformer', 'StrFeatureTransformer', 'TextClustDistTransformer'
-        , 'TextClustTETransformer', 'TextLinModelTransformer', 'TextTransformer', 'TruncSVDNumTransformer'
-        , 'WeightOfEvidenceTransformer']
 
-    # Any Installed Custom Transformers you don't want to test
-    transformers_custom_nontesting = ['MyLogTransformer']
+# Load data if it's not already on the system, return the data set key
+# TODO: re-write the already uploaded check to account for numpy warning of type mismatch
+def load_data(print_details=True):
 
-    # A list of all transformers we don't want in our experiment
-    all_nontest_transformers = transformers_noncustom + transformers_custom_nontesting
-
-    # Step Zero: Connect to Driverless AI
-    h2oai = connect()
-
-    # Step One: Load Data Set
-
-    # Get all data sets that are already loaded into DAI
     all_data_sets = h2oai.list_datasets(0, 100, include_inactive=True).datasets
     all_data_sets = pd.DataFrame({
         'key': list(map(lambda x: x.key, all_data_sets))
         , 'name': list(map(lambda x: x.name, all_data_sets))})
 
-    print("PRE-LOADED DATASETS:")
-    print(all_data_sets)
-
-    # check if data was pre-loaded - if so use that data set - if not load data
-    # this means that if your base data was changed you will need to delete the data from the UI
-    # or remove it using h2oai.delete_dataset(key)
     if data_file_name in all_data_sets['name'].values:
-        print("Data already loaded ", data_file_name)
-        data_key = all_data_sets[all_data_sets["name"] == data_file_name]["key"][0]
-        # data_load_job = h2oai.get_dataset_job(data_key).entity
+        # [0] is used so we get a sting and not a pandas.core.series.Series
+        dai_dataset_key = all_data_sets[all_data_sets["name"] == data_file_name]["key"][0]
     else:
-        print("Loading file ", data_file_name)
-        data_load_job = h2oai.upload_dataset_sync(data_file_name)
-        data_key = data_load_job.key
+        data_load_job = h2oai.upload_dataset_sync(data_file_location)
+        dai_dataset_key = data_load_job.key
 
-    # Step Two: Load custom transformer
-    # probably not good to just upload every time
-    # no function to delete from python, only from ssh-ing in
-    # rm tmp/contrib/transformers/[function]_randomletters_content.py
+    if print_details:
+        print("LOAD DATA: ")
+        print("\tExisting data on the system:")
+        print(all_data_sets)
+        print()
+        print("\tData key for Experiment: ", dai_dataset_key)
+        print()
 
-    print("Uploading Transformer ", transformer_file_name)
-    my_transformer = h2oai.upload_custom_recipe_sync(transformer_file_name)
+    return dai_dataset_key
 
-    # returns true or false - exit if fails - check DAI UI for error message (make new experiment & upload)
-    if my_transformer:
-        print("Transformer uploaded successfully")
-    else:
-        print("Transformer uploaded failed, exiting program.")
-        sys.exit()
 
-    # Step Three: Run experiment (and related tasks)
-    print("Starting Experiment")
+# Run an experiment on the fastest settings with only the transformer we are using
+# TODO: test what happens if transformer is included with overrides but has hardcoded settings above 1/1/10
+# TODO: currently assumes classification problem
+# TODO: download logs if it fails
+# TODO: speed up by turning off shift detection, python scoring pipeline etc. etc.
+def run_test_experiment(dai_dataset_key, print_details=True):
+
+    if print_details:
+        print("RUN TEST EXPERIMENT:")
+        print("\tStarting Experiment")
+
     experiment = h2oai.start_experiment_sync(
-        dataset_key=data_key
+        dataset_key=dai_dataset_key
         , target_col=y
         , is_classification=True
         , accuracy=1
@@ -119,56 +121,71 @@ def test_debug_pyclient():
         , interpretability=10
         , scorer="F1"
         , score_f_name=None
-        , config_overrides="""
-                                    feature_brain_level=0
-                                    excluded_transformers={dont_use}
-                                    """.format(dont_use=all_nontest_transformers)
+        , config_overrides="included_transformers=['" + transformer_name + "']"
     )
 
-    # if you have a pre-run experiment you want to examine
-    # experiment = h2oai.get_model_job("lomotare").entity
+    if print_details:
+        print("\tExperiment key: ", experiment.key)
+        print()
 
-    # Step Four: Check the transformation was used
+    return experiment.key
 
-    summary_path = h2oai.download(src_path=experiment.summary_path, dest_dir=".")
+
+# Print all features of the final model by downloading the experiment summary
+# TODO: should error or warning if our BYOR Transformer isn't there - in theory it should always be the only feature
+def print_model_features(dai_experiment_key, delete_downloads=True):
+    experiment = h2oai.get_model_job(dai_experiment_key).entity
+
+    summary_path = h2oai.download(src_path=experiment.summary_path, dest_dir=download_file_location)
     dir_path = "h2oai_experiment_summary_" + experiment.key
-    import zipfile
+
     with zipfile.ZipFile(summary_path, 'r') as z:
         z.extractall(dir_path)
 
-    # View Features, hopefully your transformer is here!
-    features = pd.read_table(dir_path + "/features.txt", sep=',', skipinitialspace=True)
+    features = pd.read_csv(dir_path + "/features.txt", sep=',', skipinitialspace=True)
+    print("PRINT MODEL FEATURES:")
     print(features)
+    print()
 
-    # Step Five: Transform data and ensure it looks as expected
-    transform = h2oai.fit_transform_batch_sync(model_key=experiment.key
-                                               , training_dataset_key=data_key
-                                               , validation_dataset_key=None
+    # Delete downloaded files
+    if delete_downloads:
+        os.remove(summary_path)
+        shutil.rmtree(dir_path)
+
+
+# Print the results of the BYOR transformer on your dataset
+# TODO: have only tested using the same dataset in train and validaiont on non-validated needed transformers
+def print_transformed_data(dai_experiment_key, dai_dataset_key, delete_downloads=True):
+
+    # We train and validate on the same data to get back all of th rows in the right order in transform_train
+    transform = h2oai.fit_transform_batch_sync(model_key=dai_experiment_key
+                                               , training_dataset_key=dai_dataset_key
+                                               , validation_dataset_key=dai_dataset_key
                                                , test_dataset_key=None
-                                               , validation_split_fraction=0.25
+                                               , validation_split_fraction=0
                                                , seed=1234
                                                , fold_column=None)
 
-    # Download the training and validation transformed data
-    transform_train_path = h2oai.download(src_path=transform.training_output_csv_path, dest_dir=".")
-    transform_validate_path = h2oai.download(src_path=transform.validation_output_csv_path, dest_dir=".")
+    transform_train_path = h2oai.download(src_path=transform.training_output_csv_path, dest_dir=download_file_location)
 
-    transform_train = pd.read_table(transform_train_path, sep=',', skipinitialspace=True)
-    transform_validate = pd.read_table(transform_validate_path, sep=',', skipinitialspace=True)
+    transform_train = pd.read_csv(transform_train_path, sep=',', skipinitialspace=True)
 
-    print(transform_train.head())
-    print(transform_validate.head())
+    print("PRINT TRANSFORMED DATA:")
+    print(transform_train.head(10))
+    print()
 
-    # Step Six: Join back to your training data
-
-    # Step Seven: Clean up
-    os.remove(summary_path)
-    os.remove(transform_train_path)
-    os.remove(transform_validate_path)
-    shutil.rmtree(dir_path)
+    # Delete downloaded files
+    if delete_downloads:
+        os.remove(transform_train_path)
 
 
 # run this test with `pytest -s how_to_test_from_py_client.py` or `python how_to_test_from_py_client.py`
 if __name__ == '__main__':
-    print_system_custom_transformers()
-    test_debug_pyclient()
+    dai_transformer_list = get_transformers()
+    load_transformer()
+
+    data_key = load_data()
+    experiment_key = run_test_experiment(data_key)
+
+    print_model_features(experiment_key)
+    print_transformed_data(experiment_key, data_key)
